Working on tests.
