+++ src/main/webapp/resources/js/apis/remote-api/remote-api.js
@@ -2,6 +2,6 @@ import axios from "axios";
 
 export function checkConnectionStatus({ id }) {
   return axios
-    .get(`remote_api/status/${id}`)
+    .get(`ajax/remote_api/status/${id}`)
     .then(({ data }) => data === "valid_token");
 }
